% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FPOPDp}
\alias{FPOPDp}
\title{FPOPDp}
\usage{
FPOPDp(data, penalty, type)
}
\arguments{
\item{data}{is a matrix of data(p-rows x n-columns).}

\item{penalty}{is a value of penalty (a non-negative real number).}

\item{type}{is a value defining the  type of geometry for FPOP-pruning: type=1: ("intersection" of sets), approximation - rectangle; type=2:("intersection" of sets)"minus"("union" of sets), approximation - rectangle; type=3: (last disk)"minus"("union" of sets), approximation - disk.}
}
\value{
a list of  elements  = (changepoints, means, globalCost).                    
 
\describe{
\item{\code{chpts}}{is the changepoint vector that gives the last index of each segment for the p-variate time series.}
\item{\code{means}}{is the list of successive means for the p-variate time series.}
\item{\code{globalCost}}{is a number equal to the global cost.}
}
}
\description{
Detection changepoints using the Functional Pruning Optimal Partitioning method (FPOP) in p-variate time series in a p-variable time series of length n.
}
\examples{
FPOPDp(data = data_genDp(p = 3, n = 100, chpts = 100, means = matrix(c (0, 0, 0), nrow = 3), noise = 1), penalty = 2*log(100), type = 2) 
}
