% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPOPapprox2D.R
\name{data_genDp}
\alias{data_genDp}
\title{data_genDp}
\usage{
data_genDp(p, n, chpts, means, noise = 1)
}
\arguments{
\item{p}{parameter of dimension.}

\item{n}{number of data point.}

\item{chpts}{a vector of increasing changepoint indices (last element is always n!).}

\item{means}{matrix of successive means for data.}

\item{noise}{standard deviation of an additional normal noise.}
}
\value{
matrix of data of dimension p x n with a given values of means by the segmentation.
}
\description{
Generation of data of dimension 2 with a given values of means and changepoints
}
\examples{
Data =  data_genDp(p = 3, n = 100, chpts = c(100), means = matrix(c (0, 0, 0), nrow = 3), noise = 1) 
}
